hw02: fix review issues
