+++ homework02/include/matches/returncodes.h
@@ -13,6 +13,8 @@ enum MATCHES_ERROR_CODES {
     MATCHES_OPENFILE_ERROR,
     MATCHES_ISNOTFILE_ERROR,
     MATCHES_MMAP_ERROR,
+    MATCHES_MUNMAP_ERROR,
+    MATCHES_FORK_ERROR,
     MATCHES_CHILD_ERROR,
     MATCHES_WAIT_ERROR,
     MATCHES_3RDPTY_ERROR,
