+++ homework02/lib/collections/include/collections/string.h
@@ -6,8 +6,6 @@ typedef struct string {
 } string_t;
 
 int string_init(string_t* self, const char* str);
-int string_deinit(string_t* self);
-
-const char* string_raw(string_t* self);
+void string_deinit(string_t* self);
 
 #endif  // PARKMAIL_HW_COLLECTIONS_STRING_H_
